fix: arrumei os services em geral e agr atualiza o estoque certinho

+++ src/services/clientService.ts
@@ -1,3 +1,5 @@
+import { EntityInUseError } from "../errors/EntityInUseError";
+import { NotFoundError } from "../errors/NotFoundError";
 import { PrismaClient } from "../generated/prisma";
 
 const prisma = new PrismaClient();
@@ -13,22 +15,46 @@ async function getClients() {
 }
 
 async function getById(id: number) {
+
+    const client = await prisma.client.findUnique({ where: { id } });
+
+    if (!client) {
+        throw new NotFoundError("Cliente não encontrado.");
+    }
+
     return await prisma.client.findUnique({
         where: { id },
     });
 }
 
 async function update(id: number, clientDto: Partial<any>) {
+    const client = await prisma.client.findUnique({ where: { id } });
+
+    if (!client) {
+        throw new NotFoundError("Cliente não encontrado.");
+    }
+
     return await prisma.client.update({
         where: { id },
         data: clientDto
-    }).catch(() => null);
+    });
 }
 
 async function deleteById(id: number) {
+
+    const client = await prisma.client.findUnique({ where: { id }, include: { orders: true } });
+
+    if (!client) {
+        throw new NotFoundError("Cliente não encontrado.");
+    }
+
+    if (client.orders.length > 0) {
+        throw new EntityInUseError("Cliente não pode ser deletado pois está associado a um ou mais registros.")
+    }
+
     return await prisma.client.delete({
         where: { id },
-    }).catch(() => null);
+    });
 }
 
 export { createClient, getClients, getById, update, deleteById };
